devclean/requirements.txt
# DevClean dependencies
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Scheduling
apscheduler>=3.10.0
pytz>=2024.1

# AI Orchestration
openclaw>=2026.2.28

# Optional: For OpenRouter API fallback
# (Set OPENROUTER_API_KEY in .env to use)